# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Goa Houseboat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface Package {
  id: number;
  name: string;
  description: string;
  pricePerNight: string;
  capacity: number;
  inclusions: string[];
  images: string[];
  isActive: boolean;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePackageBody {
  name: string;
  description: string;
  pricePerNight: string;
  capacity: number;
  inclusions?: string[];
  images?: string[];
  sortOrder?: number;
}

export interface UpdatePackageBody {
  name?: string;
  description?: string;
  pricePerNight?: string;
  capacity?: number;
  inclusions?: string[];
  images?: string[];
  isActive?: boolean;
  sortOrder?: number;
}

export interface Activity {
  id: number;
  name: string;
  description: string;
  icon: string;
  /** @nullable */
  image: string | null;
  isActive: boolean;
  sortOrder: number;
}

export interface CreateActivityBody {
  name: string;
  description: string;
  icon: string;
  image?: string;
  sortOrder?: number;
}

export interface UpdateActivityBody {
  name?: string;
  description?: string;
  icon?: string;
  image?: string;
  isActive?: boolean;
  sortOrder?: number;
}

export interface GalleryImage {
  id: number;
  url: string;
  /** @nullable */
  caption: string | null;
  category: string;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryImageBody {
  url: string;
  caption?: string;
  category: string;
  sortOrder?: number;
}

export type BlogPostStatus =
  (typeof BlogPostStatus)[keyof typeof BlogPostStatus];

export const BlogPostStatus = {
  pending: "pending",
  published: "published",
  rejected: "rejected",
} as const;

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  excerpt: string | null;
  authorName: string;
  /** @nullable */
  authorEmail: string | null;
  images: string[];
  status: BlogPostStatus;
  isAdminPost: boolean;
  /** @nullable */
  metaTitle: string | null;
  /** @nullable */
  metaDescription: string | null;
  /** @nullable */
  publishedAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostList {
  posts: BlogPost[];
  total: number;
  page: number;
  limit: number;
}

export interface CreateBlogPostBody {
  title: string;
  content: string;
  excerpt?: string;
  authorName: string;
  authorEmail?: string;
  images?: string[];
  otpToken?: string;
  metaTitle?: string;
  metaDescription?: string;
}

export type UpdateBlogPostBodyStatus =
  (typeof UpdateBlogPostBodyStatus)[keyof typeof UpdateBlogPostBodyStatus];

export const UpdateBlogPostBodyStatus = {
  pending: "pending",
  published: "published",
  rejected: "rejected",
} as const;

export interface UpdateBlogPostBody {
  title?: string;
  content?: string;
  excerpt?: string;
  images?: string[];
  status?: UpdateBlogPostBodyStatus;
  metaTitle?: string;
  metaDescription?: string;
}

export interface SendOtpBody {
  email: string;
}

export interface VerifyOtpBody {
  email: string;
  otp: string;
}

export interface OtpVerifyResponse {
  token: string;
  message: string;
}

export interface Settings {
  id: number;
  siteName: string;
  tagline: string;
  heroTitle: string;
  heroSubtitle: string;
  heroImage: string;
  whatsappNumber: string;
  inquiryEmail: string;
  /** @nullable */
  socialInstagram: string | null;
  /** @nullable */
  socialFacebook: string | null;
  /** @nullable */
  socialYoutube: string | null;
  /** @nullable */
  trailVideoUrl: string | null;
  trailTitle: string;
  trailDescription: string;
  aboutText: string;
  aboutImages: string[];
  /** @nullable */
  locationMapUrl?: string | null;
  showChatWidget?: string;
  chatWidgetColor?: string;
  chatWidgetAlignment?: string;
  showWhatsappButton?: string;
  updatedAt: string;
}

export interface UpdateSettingsBody {
  siteName?: string;
  tagline?: string;
  heroTitle?: string;
  heroSubtitle?: string;
  heroImage?: string;
  whatsappNumber?: string;
  inquiryEmail?: string;
  socialInstagram?: string;
  socialFacebook?: string;
  socialYoutube?: string;
  trailVideoUrl?: string;
  trailTitle?: string;
  trailDescription?: string;
  aboutText?: string;
  aboutImages?: string[];
  locationMapUrl?: string;
  showChatWidget?: string;
  chatWidgetColor?: string;
  chatWidgetAlignment?: string;
  showWhatsappButton?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
  role: string;
}

export interface LoginResponse {
  user: AdminUser;
  message: string;
}

export interface InquiryBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  checkIn?: string | null;
  /** @nullable */
  checkOut?: string | null;
  /** @nullable */
  guests?: number | null;
  message: string;
}

export type ListGalleryParams = {
  category?: string;
};

export type ListBlogPostsParams = {
  page?: number;
  limit?: number;
};