# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Goa Houseboat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List all activities
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
  /activities/{id}:
    patch:
      operationId: updateActivity
      tags: [activities]
      summary: Update activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateActivityBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
    delete:
      operationId: deleteActivity
      tags: [activities]
      summary: Delete activity (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user session
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - name
        - description
        - icon
        - image
        - isActive
        - sortOrder
    CreateActivityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        image:
          type: string
        sortOrder:
          type: integer
      required:
        - name
        - description
        - icon
    UpdateActivityBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        image:
          type: string
        isActive:
          type: boolean
        sortOrder:
          type: integer
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/AdminUser"
        message:
          type: string
      required:
        - user
        - message
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
      required:
        - id
        - username
        - role
...