# lib/api-zod/src/generated/types/createActivityBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Goa Houseboat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateActivityBody {
  name: string;
  description: string;
  icon: string;
  image?: string;
  sortOrder?: number;
}